requirements.txt
alphamini
protobuf==3.20.3
websockets==10.3  # <-- ДОДАЄМО ЦЕ